motor codez

+++ code/timer.h
@@ -1,9 +1,11 @@
 #ifndef _TIMER_H
 #define _TIMER_H
 
+#include <stdint.h>
+
 void initTimer();
-unsigned long millis();
-unsigned long micros();
+uint32_t millis();
+uint32_t micros();
 
 #endif
